Fixes to tests
