+++ .github/workflows/package.yml
@@ -37,17 +37,3 @@ jobs:
           pip install -e .
           pip install pytest
           python3 -m pytest tests/package/*.py
-  extra-models-package-install-check:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Install pip
-        run: python3 -m pip install --upgrade pip
-      - uses: actions/setup-python@v4.7.1
-        with:
-          python-version: '3.10'
-      - name: Install package and test import
-        run: |
-          pip install -e ".[models]"
-          pip install pytest
-          python3 -m pytest tests/package/extras/models.py
