+++ pyproject.toml
@@ -139,6 +139,7 @@ lightning = "^2.1.0"
 
 [tool.poetry.extras]
 torch = ["torch"]
+torchvision = ["torchvision"]
 torchxrayvision = ["torch", "torchxrayvision"]
 monai = ["torch", "monai"]
 xgboost = ["xgboost"]
